requirements.txt
djangorestframework
Django==2.0
pytz==2019.2
